use old versions

+++ dockerstart.sh
@@ -4,22 +4,39 @@
 IMAGE_NAME="payme_image"
 CONTAINER_NAME="payme_container"
 
-# Check if the container is already running, and remove it if it is
-if [ $(docker ps -q -f name=^/${CONTAINER_NAME}$) ]; then
+# Function to check if a docker container exists
+container_exists() {
+  docker ps -a -q -f name=^/${CONTAINER_NAME}$
+}
+
+# Function to check if a docker container is running
+container_is_running() {
+  docker ps -q -f name=^/${CONTAINER_NAME}$
+}
+
+# Stop and remove container if it is running
+if container_is_running; then
   echo "Stopping and removing existing container..."
   docker stop "${CONTAINER_NAME}"
   docker rm "${CONTAINER_NAME}"
   echo "Container removed."
+elif container_exists; then
+  # Remove container if it exists but is not running
+  echo "Removing existing container..."
+  docker rm "${CONTAINER_NAME}"
+  echo "Container removed."
 fi
-
-docker rm "${CONTAINER_NAME}"
 
 # Build the Docker image
 echo "Building Docker image..."
 docker build -t "${IMAGE_NAME}" .
 
 # Run the container in detached mode
+# Mount the entire project directory as a volume
 echo "Running container in detached mode..."
-docker run -d -p 8501:8501 --name "${CONTAINER_NAME}" "${IMAGE_NAME}"
+docker run -d -p 8501:8501 \
+  -v $(pwd):/home/projects/payme \
+  --name "${CONTAINER_NAME}" \
+  "${IMAGE_NAME}"
 
 echo "Container ${CONTAINER_NAME} is up and running."
